feat(assistant): make assistant link rendering configurable via renderLink
